refactor(compiler): move `findPathsInPacks` to `path`
